fixed numpy deprecation in generate_bins

+++ polychrom/polymer_analyses.py
@@ -130,7 +130,7 @@ def smart_contacts(data, cutoff=1.7, min_cutoff=2.1, percent_func=lambda x:1/x):
 def generate_bins(N, start=4, bins_per_order_magn=10):
     lstart = np.log10(start)
     lend = np.log10(N - 1) + 1e-6
-    num = np.ceil((lend - lstart) * bins_per_order_magn)
+    num = int(np.ceil((lend - lstart) * bins_per_order_magn))
     bins = np.unique(np.logspace(lstart, lend, dtype=int, num=max(num,0)))
     if len(bins) > 0:
         assert bins[-1] == N - 1
